python/pyarmnn/examples/requirements.txt
requests>=2.23.0
urllib3>=1.25.8
Pillow>=6.1.0
numpy>=1.18.1
pyarmnn>=19.8.0
